+++ deploy/scripts/deploy-zeroex-v2.js
@@ -1,11 +1,11 @@
 const {nab, call, send} = require('../utils/deploy-contract');
-const { assetDataUtils } = require('@0x/order-utils');
+const { assetDataUtils } = require('@0x/order-utils-v2');
 
 const zeroAddress = '0x0000000000000000000000000000000000000000'; // TODO: import from util
 
 const main = async input => {
-  const exchange = await nab('ZeroExV2Exchange', [], input.zeroex.addr);
-  const erc20Proxy = await nab('ERC20Proxy', [], input.zeroex.addr);
+  const exchange = await nab('ZeroExV2Exchange', [], input.zeroExV2.addr);
+  const erc20Proxy = await nab('ZeroExV2ERC20Proxy', [], input.zeroExV2.addr);
 
   const alreadyAuth = await call(erc20Proxy, 'authorized', [exchange.options.address]);
   if (!alreadyAuth) {
@@ -26,7 +26,7 @@ const main = async input => {
 
   return {
     "ZeroExV2Exchange": exchange,
-    "ERC20Proxy": erc20Proxy
+    "ZeroExV2ERC20Proxy": erc20Proxy
   };
 }
